requirements.txt
openai-whisper==20231117
torch==2.2.0
python-dotenv==1.0.1
tqdm==4.66.2
torchaudio==2.2.0
resemblyzer==0.1.4
scikit-learn==1.4.0
numpy==1.26.4